+++ apps/backend/src/libs/plugins/authorization/helpers/helpers.ts
@@ -1,1 +1,1 @@
-export { checkIsWhiteRoute } from "./check-is-white-route.helper.js";
+export { checkIsWhiteRoute } from "./check-is-white-route/check-is-white-route.helper.js";