+++ apps/frontend/src/libs/modules/http/libs/enums/enums.ts
@@ -1,1 +1,2 @@
+export { ExceptionName } from "./exception-name.enum.js";
 export { HTTPCode, HTTPHeader } from "@git-fit/shared";