+++ apps/frontend/src/libs/modules/http/libs/enums/exception-name.enum.ts
@@ -1,0 +1,5 @@
+const ExceptionName = {
+	UNAUTHORIZED: "Unauthorized",
+} as const;
+
+export { ExceptionName };