+++ apps/frontend/src/pages/access-management/libs/components/group-users-table/libs/helpers/helpers.ts
@@ -1,1 +1,1 @@
-export { getGroupUsersColumns } from "./get-group-users-columns.helper.js";
+export { getGroupUsersColumns } from "./get-group-users-columns/get-group-users-columns.helper.js";