+++ apps/frontend/src/pages/project/project.tsx
@@ -1,7 +1,9 @@
 import {
 	Breadcrumbs,
 	Button,
+	ConfirmationModal,
 	Modal,
+	Navigate,
 	PageLayout,
 } from "~/libs/components/components.js";
 import { AppRoute, DataStatus, PermissionKey } from "~/libs/enums/enums.js";
@@ -49,6 +51,7 @@ const Project = (): JSX.Element => {
 		project,
 		projectContributors,
 		projectContributorsStatus,
+		projectDeleteStatus,
 		projectPatchStatus,
 		projectStatus,
 	} = useAppSelector(({ projects }) => projects);
@@ -63,6 +66,12 @@ const Project = (): JSX.Element => {
 		isOpened: isEditModalOpen,
 		onClose: handleEditModalClose,
 		onOpen: handleEditModalOpen,
+	} = useModal();
+
+	const {
+		isOpened: isDeleteModalOpen,
+		onClose: handleDeleteModalClose,
+		onOpen: handleDeleteModalOpen,
 	} = useModal();
 
 	const {
@@ -104,6 +113,10 @@ const Project = (): JSX.Element => {
 		handleEditModalOpen();
 	}, [handleEditModalOpen]);
 
+	const handleDeleteProject = useCallback(() => {
+		handleDeleteModalOpen();
+	}, [handleDeleteModalOpen]);
+
 	const handleProjectEditSubmit = useCallback(
 		(payload: ProjectPatchRequestDto) => {
 			if (project) {
@@ -112,6 +125,12 @@ const Project = (): JSX.Element => {
 		},
 		[dispatch, project],
 	);
+
+	const handleProjectDeleteConfirm = useCallback(() => {
+		if (project) {
+			void dispatch(projectActions.deleteById(project.id));
+		}
+	}, [project, dispatch]);
 
 	const handleEditContributor = useCallback(
 		(contributorId: number) => {
@@ -191,6 +210,15 @@ const Project = (): JSX.Element => {
 		return <NotFound />;
 	}
 
+	const isProjectDeleted =
+		hasProject &&
+		project.id === Number(projectId) &&
+		projectDeleteStatus === DataStatus.FULFILLED;
+
+	if (isProjectDeleted) {
+		return <Navigate to={AppRoute.PROJECTS} />;
+	}
+
 	return (
 		<PageLayout isLoading={isLoading}>
 			{hasProject && (
@@ -209,6 +237,7 @@ const Project = (): JSX.Element => {
 							<h1 className={styles["title"]}>{project.name}</h1>
 
 							<ProjectDetailsMenu
+								onDelete={handleDeleteProject}
 								onEdit={handleEditProject}
 								projectId={project.id}
 								userPermissions={userPermissions}
@@ -285,6 +314,13 @@ const Project = (): JSX.Element => {
 						isOpened={isSetupAnalyticsModalOpened}
 						onClose={onSetupAnalyticsModalClose}
 						project={project}
+					/>
+
+					<ConfirmationModal
+						content="The project will be deleted. This action cannot be undone. Click 'Confirm' to proceed."
+						isOpened={isDeleteModalOpen}
+						onClose={handleDeleteModalClose}
+						onConfirm={handleProjectDeleteConfirm}
 					/>
 				</>
 			)}