+++ packages/shared/src/modules/activity-logs/libs/types/types.ts
@@ -1,5 +1,8 @@
 export { type ActivityLogCreateItemRequestDto } from "./activity-log-create-item-request-dto.type.js";
 export { type ActivityLogCreateItemResponseDto } from "./activity-log-create-item-response-dto.type.js";
 export { type ActivityLogCreateRequestDto } from "./activity-log-create-request-dto.type.js";
+export { type ActivityLogGetAllAnalyticsResponseDto } from "./activity-log-get-all-analytics-response-dto.type.js";
+export { type ActivityLogGetAllItemAnalyticsResponseDto } from "./activity-log-get-all-item-analytics-response-dto.type.js";
 export { type ActivityLogGetAllItemResponseDto } from "./activity-log-get-all-item-response-dto.type.js";
 export { type ActivityLogGetAllResponseDto } from "./activity-log-get-all-response-dto.type.js";
+export { type ActivityLogQueryParameters } from "./activity-log-query-parameters.type.js";